change `bashcompdir` to the legacy directory
The bash-completion package recommends using
`/usr/share/bash-completion/completions` as the default bash-completion
directory but the bash completion files of password store extensions
won't work if they're kept in this directory. This is because our bash
completion file will only be loaded on demand when a command called
`pass-tessen` executes, the same as the basename of our completion file.
Of course, this doesn't happen.

The legacy directory, `/etc/bash_completion.d` works as expected because
all files present inside it are loaded automatically. This directory,
however, seems to have been deprecated.

Check the following link for more details

https://github.com/scop/bash-completion#faq