refactor: get the key from a user in a second menu
The username (userkey), password, URL key (urlkey), and OTP (otp) key
are special keys that need a different treatment. The username and
password keys should be configurable like tessen. The URL key is
configurable and the value will not be typed but opened using `xdg-open`
or a web browser configuration value. All the other key value pairs will
be treated a simple data that can be copied.

If the chosen key is invalid, exit.