feat: add support for TOTP/HOTP and URLs
refactor: overhaul the action subroutine for the selected key

If 'otp' is selected, `pass-otp` is used to generate a TOTP/HOTP and it
is copied to the clipboard.
If a URL is selected, pass-tessen will look for a custom web browser
value. If not found, it'll look for `xdg-open`. If that's not found
either, an error will be shown.
All the other keys, including the username and the password are passed
to another function to be copied without any specific actions.