chore: overhaul the help menu
The help menu has been overhauled to indicate new flags and show a list
of git forges for reporting issues. The list of supported fuzzy
selection programs and the default values for the username and URL keys
are mentioned as well. The list of new optional dependencies are also
mentioned.